requirements.txt
Pillow==9.3.0
streamlit==1.17.0
streamlit-option-menu==0.3.2